requirements.txt
asgiref==3.7.2
bleach==6.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-startapi==0.0.2.dev2
django-summernote==0.8.20.0
django-treasuremap==0.3.4
djangorestframework==3.14.0
geojson==2.5.0
idna==3.4
Pillow==10.0.0
pyowm==3.3.0
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
webencodings==0.5.1
XlsxWriter==3.1.2
